Rename utils/coreutils.py to utils/misc.py
