+++ wagtail/admin/views/reports/aging_pages.py
@@ -7,7 +7,7 @@ from django.utils.translation import gettext_lazy as _
 from wagtail.admin.filters import ContentTypeFilter, WagtailFilterSet
 from wagtail.admin.widgets import AdminDateInput
 from wagtail.models import Page, PageLogEntry, UserPagePermissionsProxy, get_page_models
-from wagtail.utils.coreutils import get_content_type_label
+from wagtail.utils.misc import get_content_type_label
 
 from .base import PageReportView
