+++ wagtail/urls.py
@@ -3,7 +3,7 @@ from django.contrib.auth import views as auth_views
 from django.urls import path, re_path
 
 from wagtail import views
-from wagtail.utils.coreutils import WAGTAIL_APPEND_SLASH
+from wagtail.utils.misc import WAGTAIL_APPEND_SLASH
 
 if WAGTAIL_APPEND_SLASH:
     # If WAGTAIL_APPEND_SLASH is True (the default value), we match a
