+++ wagtail/utils/misc.py
